+++ src/config/settings.rs
@@ -327,6 +327,8 @@ pub(super) struct Guts {
     pub(super) prefix_list_stats: String,
     #[config(default = "Ignored important error: ")]
     pub(super) prefix_ignored_important_error_message: String,
+    #[config(default = "\n\tConsider reporting the error to add this tag to the list of unexpected tags to skip by default")]
+    pub(super) infix_add_unexpected_tag_suggestion: String,
     #[config(default = "\n\tFix the problem or add \"--ignore-important-errors\"(may rarely cause unexpected behaviour) to ignore")]
     pub(super) suffix_add_ignore_important_errors_suggestion: String,
 }